requirements.txt
yt_dlp
youtube_search
requests
python-telegram-bot==13.15